% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aact_query.R
\name{aact_check_connection}
\alias{aact_check_connection}
\title{Check database connection}
\usage{
aact_check_connection(con)
}
\arguments{
\item{con}{Database connection object}
}
\value{
A data frame with distinct study types
}
\description{
Check database connection
}
\examples{
\dontrun{
# Set environment variables for database credentials in .Renviron and load it
# readRenviron(".Renviron")

# Connect to the database
con <- aact_connection(Sys.getenv('user'), Sys.getenv('password'))

# Check the connection
aact_check_connection(con)
}
}
