% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aact_query.R
\name{aact_connection}
\alias{aact_connection}
\title{Connect to AACT PostgreSQL database}
\usage{
aact_connection(user, password)
}
\arguments{
\item{user}{Database username}

\item{password}{Database password}
}
\value{
A connection object to the AACT database
}
\description{
Connect to AACT PostgreSQL database
}
\examples{
\donttest{
# Set environment variables for database credentials in .Renviron and load it
# readRenviron(".Renviron")

# Connect to the database
con <- aact_connection(Sys.getenv('user'), Sys.getenv('password'))
}
}
