% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print a Welcome Message}
\usage{
hello()
}
\value{
A character string containing the welcome message.
}
\description{
This function prints a welcome message for ClinTrialX.
}
\examples{
hello()
}
