% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_info.R
\name{version_info}
\alias{version_info}
\title{Get API Version Information}
\usage{
version_info(source = "clinicaltrials.gov")
}
\arguments{
\item{source}{A character string specifying the source to query. Currently, only "clinicaltrials.gov" is supported.}
}
\value{
Prints the API version and data timestamp if successful. Otherwise, prints an error message.
}
\description{
This function retrieves version information from specified clinical trials API sources.
}
\examples{
version_info()
version_info("clinicaltrials.gov")
}
