% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_info.R
\name{version_info}
\alias{version_info}
\title{Get API Version Information}
\usage{
version_info(source = "clinicaltrials.gov")
}
\arguments{
\item{source}{A character string specifying the source to query.
Currently, "clinicaltrials.gov" and "aact" are supported.}
}
\value{
A list containing API version and data timestamp for clinicaltrials.gov,
  or NULL for aact with a message printed.
}
\description{
This function retrieves version information from specified clinical trials API sources.
}
\examples{
version_info()
version_info("clinicaltrials.gov")
version_info("aact")

}
\references{
ClinicalTrials.gov API - https://clinicaltrials.gov/api/v2/version
AACT - https://aact.ctti-clinicaltrials.org/release_notes
}
